Add GPS command handler for managing GPS info display
