+++ include/core/cli-handlers/gps-cli-handler.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void handle_gps_info(int argc, char **argv);
