+++ main/core/cli-handlers/gps-cli-handler.c
@@ -1,0 +1,42 @@
+#include "core/cli-handlers/gps-cli-handler.h"
+#include "managers/gps_manager.h"
+#include "managers/views/terminal_screen.h"
+#include "freertos/FreeRTOS.h"
+#include "freertos/task.h"
+#include <stdio.h>
+
+extern TaskHandle_t gps_info_task_handle;
+extern GPSManager g_gpsManager;
+
+void handle_gps_info(int argc, char **argv) {
+    bool stop_flag = false;
+
+    for (int i = 1; i < argc; i++) {
+        if (strcmp(argv[i], "-s") == 0) {
+            stop_flag = true;
+            break;
+        }
+    }
+
+    if (stop_flag) {
+        if (gps_info_task_handle != NULL) {
+            vTaskDelete(gps_info_task_handle);
+            gps_info_task_handle = NULL;
+            gps_manager_deinit(&g_gpsManager);
+            printf("GPS info display stopped.\n");
+            TERMINAL_VIEW_ADD_TEXT("GPS info display stopped.\n");
+        }
+    } else {
+        if (gps_info_task_handle == NULL) {
+            gps_manager_init(&g_gpsManager);
+
+            // Wait a brief moment for GPS initialization
+            vTaskDelay(pdMS_TO_TICKS(100));
+
+            // Start the info display task
+            xTaskCreate(gps_info_display_task, "gps_info", 4096, NULL, 1, &gps_info_task_handle);
+            printf("GPS info started.\n");
+            TERMINAL_VIEW_ADD_TEXT("GPS info started.\n");
+        }
+    }
+}
