+++ main/core/commandline.c
@@ -36,6 +36,7 @@
 #include "core/cli-handlers/sd-cli-handler.h"
 #include "core/cli-handlers/system-cli-handler.h"
 #include "core/cli-handlers/portal-cli-handler.h"
+#include "core/cli-handlers/gps-cli-handler.h"
 
 
 #if !defined(MAX_WIFI_CHANNEL)
@@ -1515,40 +1516,6 @@ void handle_capture(int argc, char **argv) {
     }
 #endif
 }
-
-void handle_gps_info(int argc, char **argv) {
-    bool stop_flag = false;
-
-    for (int i = 1; i < argc; i++) {
-        if (strcmp(argv[i], "-s") == 0) {
-            stop_flag = true;
-            break;
-        }
-    }
-
-    if (stop_flag) {
-        if (gps_info_task_handle != NULL) {
-            vTaskDelete(gps_info_task_handle);
-            gps_info_task_handle = NULL;
-            gps_manager_deinit(&g_gpsManager);
-            printf("GPS info display stopped.\n");
-            TERMINAL_VIEW_ADD_TEXT("GPS info display stopped.\n");
-        }
-    } else {
-        if (gps_info_task_handle == NULL) {
-            gps_manager_init(&g_gpsManager);
-
-            // Wait a brief moment for GPS initialization
-            vTaskDelay(pdMS_TO_TICKS(100));
-
-            // Start the info display task
-            xTaskCreate(gps_info_display_task, "gps_info", 4096, NULL, 1, &gps_info_task_handle);
-            printf("GPS info started.\n");
-            TERMINAL_VIEW_ADD_TEXT("GPS info started.\n");
-        }
-    }
-}
-
 
 #ifndef CONFIG_IDF_TARGET_ESP32S2
 void handle_ble_wardriving(int argc, char **argv) {
